requirements.txt
fastapi
uvicorn
python-dotenv
pydantic-settings
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
pymysql
SQLAlchemy